+++ index.js
@@ -212,6 +212,10 @@ var _isISO31661Alpha = require('./lib/isISO31661Alpha2');
 
 var _isISO31661Alpha2 = _interopRequireDefault(_isISO31661Alpha);
 
+var _isISO31661Alpha3 = require('./lib/isISO31661Alpha3');
+
+var _isISO31661Alpha4 = _interopRequireDefault(_isISO31661Alpha3);
+
 var _isBase = require('./lib/isBase64');
 
 var _isBase2 = _interopRequireDefault(_isBase);
@@ -335,6 +339,7 @@ var validator = {
   isISO8601: _isISO2.default,
   isRFC3339: _isRFC2.default,
   isISO31661Alpha2: _isISO31661Alpha2.default,
+  isISO31661Alpha3: _isISO31661Alpha4.default,
   isBase64: _isBase2.default,
   isDataURI: _isDataURI2.default,
   isMimeType: _isMimeType2.default,